Refix Sheet1 (#493)
* refix sheet1

* delete unused div
